requirements.txt
Django==5.2.5
django-modeltranslation
djangorestframework==3.15.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.7
django-extensions==3.2.3
django-filter==24.1
celery==5.3.4
django-celery-results
django-celery
redis==5.0.1
django-redis==5.4.0
docutils
django-injector
adrf
pillow
# Django Channels for WebSocket support
channels==4.0.0
daphne==4.0.0
# ASGI server (alternative to Daphne)
uvicorn==0.27.1
# Production WSGI server with Uvicorn workers
gunicorn==21.2.0
pytz
# Performance testing
aiohttp==3.9.3
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.4.1
# formatter
black==24.1.1
flake8==7.0.0
isort==5.13.2
pre-commit==3.6.0